cli: move check-store to own file
Release note: None
